\name{addNumStrings}
\alias{addNumStrings}
\title{
Add numeric strings.
}
\description{
A test function for adding numeric strings.
}
\usage{
addNumStrings(s)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{s}{
A vector of numeric strings.
}
}
\value{
A numeric string.
}

\examples{
addNumStrings(c("1.2345345", "-0.34534", "3.1415900"))
}



